Glacius now builds & runs
StatusServer doesn't work on Linux, rest is untested. Messages.hpp is just copied from Tolcl source tree. (it's the right one, but this is the wrong way to do it)

Bringing back MySQL support would require resurrecting BlueShine first.

+++ src/Glacius.cpp
@@ -49,7 +49,7 @@ static void run( const String& configFile )
 
         StatusServer* status = 0;
 
-        if ( confGlobal->getOption( "StatusServer/enabled", true ) )
+        if ( confGlobal->getOption( "StatusServer/enabled", true ).toInt() )
         {
             printf( "## starting Status Server\n" );
             status = new StatusServer();
